requirements.txt
selenium==4.16.0
pytest==7.4.3
allure-pytest==2.9.45
requests==2.26.0
pytest-xdist==2.2.1